requirements.txt
# RiskRaider NFT Analysis Discord Bot - Core Dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0

# NFT & Blockchain APIs
requests>=2.31.0
httpx>=0.26.0
web3>=6.15.1

# Data Processing & Analytics
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database & Storage
asyncpg>=0.29.0
redis>=5.0.0

# Utility & Helper Libraries
python-dateutil>=2.8.0
pytz>=2023.3
humanize>=4.8.0
emoji>=2.8.0

# Image Processing & Visualization
Pillow>=10.0.0
qrcode[pil]>=7.4.0

# API Rate Limiting & Caching
aiocache>=0.12.0
asyncio-throttle>=1.0.0

# Logging & Monitoring
structlog>=23.1.0
sentry-sdk>=1.39.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: AI/ML for advanced analysis
scikit-learn>=1.3.0
joblib>=1.3.0